move chainmgr.BlockPool to xshard
