+++ core/chainmgr/message/shardblock_info.go
@@ -19,20 +19,12 @@
 package message
 
 import (
-	"bytes"
 	"fmt"
 	"io"
 
 	"github.com/ontio/ontology/common"
-	"github.com/ontio/ontology/common/log"
 	"github.com/ontio/ontology/core/types"
 	"github.com/ontio/ontology/events/message"
-)
-
-const (
-	ShardBlockNew = iota
-	ShardBlockReceived
-	ShardBlockProcessed
 )
 
 //
@@ -60,7 +52,6 @@ func (this *ShardBlockTx) Deserialization(source *common.ZeroCopySource) error {
 type ShardBlockInfo struct {
 	FromShardID common.ShardID                   `json:"from_shard_id"`
 	Height      uint32                           `json:"height"`
-	State       uint                             `json:"state"`
 	Block       *types.Block                     `json:"block"`
 	ShardTxs    map[common.ShardID]*ShardBlockTx `json:"shard_txs"` // indexed by ToShardID
 }
@@ -68,7 +59,6 @@ type ShardBlockInfo struct {
 func (this *ShardBlockInfo) Serialization(sink *common.ZeroCopySink) error {
 	sink.WriteUint64(this.FromShardID.ToUint64())
 	sink.WriteUint32(this.Height)
-	sink.WriteUint64(uint64(this.State))
 	this.Block.Serialization(sink)
 	return nil
 }
@@ -84,11 +74,9 @@ func (this *ShardBlockInfo) Deserialization(source *common.ZeroCopySource) error
 	}
 	this.FromShardID = id
 	this.Height, eof = source.NextUint32()
-	state, eof := source.NextUint64()
 	if eof {
 		return io.ErrUnexpectedEOF
 	}
-	this.State = uint(state)
 	this.Block = &types.Block{}
 	if err := this.Block.Deserialization(source); err != nil {
 		return fmt.Errorf("deserialization: read header failed, err: %s", err)
@@ -103,78 +91,5 @@ func (this *ShardBlockInfo) Deserialization(source *common.ZeroCopySource) error
 			return fmt.Errorf("deserialization: read event failed, index %d, err: %s", i, err)
 		}
 	}
-	return nil
-}
-
-////////////////////////////////////
-//
-//  shard block pool
-//
-////////////////////////////////////
-
-type ShardBlockMap map[uint32]*ShardBlockInfo // indexed by BlockHeight
-
-type ShardBlockPool struct {
-	Shards      map[common.ShardID]ShardBlockMap // indexed by FromShardID
-	MaxBlockCap uint32
-}
-
-func NewShardBlockPool(historyCap uint32) *ShardBlockPool {
-	return &ShardBlockPool{
-		Shards:      make(map[common.ShardID]ShardBlockMap),
-		MaxBlockCap: historyCap,
-	}
-}
-
-func (pool *ShardBlockPool) GetBlockInfo(shardID common.ShardID, height uint32) *ShardBlockInfo {
-	if m, present := pool.Shards[shardID]; present && m != nil {
-		return m[height]
-	}
-	return nil
-}
-
-func (pool *ShardBlockPool) AddBlockInfo(blkInfo *ShardBlockInfo) error {
-	if _, present := pool.Shards[blkInfo.FromShardID]; !present {
-		pool.Shards[blkInfo.FromShardID] = make(ShardBlockMap)
-	}
-
-	m := pool.Shards[blkInfo.FromShardID]
-	if m == nil {
-		return fmt.Errorf("add shard block, nil map")
-	}
-	if blk, present := m[blkInfo.Height]; present {
-		if blk.State != ShardBlockNew {
-			return fmt.Errorf("add shard block, new block on block state %d", blk.State)
-		}
-		hdr := blk.Block.Header
-		if hdr != nil && bytes.Compare(hdr.BlockRoot[:], blkInfo.Block.Header.BlockRoot[:]) == 0 {
-			return fmt.Errorf("add shard block, dup blk")
-		}
-	}
-
-	log.Infof("chainmgr AddBlock from shard %d, block %d", blkInfo.FromShardID, blkInfo.Height)
-	m[blkInfo.Height] = blkInfo
-
-	// if too much block cached in map, drop old blocks
-	if uint32(len(m)) < pool.MaxBlockCap {
-		return nil
-	}
-	h := blkInfo.Height
-	for _, blk := range m {
-		if blk.Height > h {
-			h = blk.Height
-		}
-	}
-
-	toDrop := make([]uint32, 0)
-	for _, blk := range m {
-		if blk.Height < h-uint32(pool.MaxBlockCap) {
-			toDrop = append(toDrop, blk.Height)
-		}
-	}
-	for _, blkHeight := range toDrop {
-		delete(m, blkHeight)
-	}
-
 	return nil
 }
