+++ core/chainmgr/xshard/shard_state_query.go
@@ -25,16 +25,11 @@ import (
 	"github.com/ontio/ontology/common"
 	"github.com/ontio/ontology/core/ledger"
 	sComm "github.com/ontio/ontology/core/store/common"
-	"github.com/ontio/ontology/core/types"
 	"github.com/ontio/ontology/smartcontract/service/native/shard_stake"
 	"github.com/ontio/ontology/smartcontract/service/native/shardmgmt"
 	"github.com/ontio/ontology/smartcontract/service/native/shardmgmt/states"
 	"github.com/ontio/ontology/smartcontract/service/native/utils"
 )
-
-func GetShardTxsByParentHeight(start, end uint32) map[common.ShardID][]*types.Transaction {
-	return nil
-}
 
 func GetShardView(lgr *ledger.Ledger, shardID common.ShardID) (*utils.ChangeView, error) {
 	if lgr == nil {
